convert to OOP format with new classes Ui, Storage, Parser, TaskList
