+++ README.md
@@ -1,4 +1,4 @@
-# Goon project template
+# goon.Goon project template
 
 This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
