+++ data/tasks.txt
@@ -6,3 +6,4 @@ T | 1 | join sports club
 T | 0 | asdfasdlkj
 D | 1 | another asn /by 2026-08-02
 D | 0 | lastdead /by 2025-06-01
+T | 0 | testing
