+++ src/main/java/goon/GoonException.java
@@ -1,3 +1,5 @@
+package goon;
+
 public class GoonException extends Exception {
     public GoonException() {}
