+++ src/main/java/goon/Parser.java
@@ -1,0 +1,4 @@
+package goon;
+
+public class Parser {
+}
