+++ src/main/java/goon/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package goon.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
