+++ src/main/java/goon/tasks/Event.java
@@ -1,3 +1,5 @@
+package goon.tasks;
+
 public class Event extends Task {
     protected String from;
     protected String to;
