+++ src/main/java/goon/tasks/Task.java
@@ -1,3 +1,5 @@
+package goon.tasks;
+
 public class Task {
     protected String description;
     protected boolean isDone;
